+++ verik-compiler/src/main/kotlin/io/verik/compiler/specialize/SpecializerCopier.kt
@@ -161,7 +161,6 @@ object SpecializerCopier {
             declarations = ArrayList(declarations),
             primaryConstructor = primaryConstructor,
             isEnum = cls.isEnum,
-            isAbstract = cls.isAbstract,
             isObject = cls.isObject
         )
         specializeContext.register(cls, typeArguments, copiedClass)
@@ -201,7 +200,6 @@ object SpecializerCopier {
             body = body,
             valueParameters = ArrayList(valueParameters),
             typeParameters = ArrayList(typeParameters),
-            isAbstract = function.isAbstract,
             isOverride = function.isOverride
         )
         specializeContext.register(function, typeArguments, copiedFunction)